+++ chatlas/__init__.py
@@ -14,6 +14,7 @@ from ._provider_github import ChatGithub
 from ._provider_google import ChatGoogle, ChatVertex
 from ._provider_groq import ChatGroq
 from ._provider_huggingface import ChatHuggingFace
+from ._provider_mistral import ChatMistral
 from ._provider_ollama import ChatOllama
 from ._provider_openai import ChatAzureOpenAI, ChatOpenAI
 from ._provider_openrouter import ChatOpenRouter
@@ -40,6 +41,7 @@ __all__ = (
     "ChatGoogle",
     "ChatGroq",
     "ChatHuggingFace",
+    "ChatMistral",
     "ChatOllama",
     "ChatOpenAI",
     "ChatOpenRouter",